requirements.txt
Flask==2.3.3
numpy==1.24.3
matplotlib==3.7.2
mysql-connector-python==8.2.0